# server/src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: API для ЦМР
  description: API для управления пользователями, новостями, достижениями, событиями и многим другим на сайте ЦМР.
  version: 1.0.0
paths:
  /auth/sign-up:
    post:
      summary: Регистрация нового администратора
      description: Позволяет зарегистрировать нового администратора
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: zhikh23@bmstu.ru
                password:
                  type: string
                  example: timosha123
                adminKey:
                  type: string
                  example: IvanLovesTimosha
              required:
                - email
                - password
      responses:
        '200':
          description: Администратор успешно зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: number
  /auth/sign-in:
    post:
      summary: Вход в аккаунт администратора
      description: Позволяет войти в аккаунт администратора
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: timosha@mail.ru
                password:
                  type: string
                  example: Timosha123
              required:
                - email
                - password
      responses:
        '200':
          description: Администратор успешно вошёл в аккаунт
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    example: success
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: number
...